Restructure and cleaned up components and images folders
