+++ frontend/src/App.jsx
@@ -2,14 +2,14 @@ import React, { useState } from "react";
 import {
   BrowserRouter as Router, Routes, Route,
 } from "react-router-dom";
-import Header from "./components/header/Header";
 import CourseSelector from "./pages/CourseSelector/main";
 import DegreeWizard from "./pages/DegreeWizard/main";
 import ProgressionChecker from "./pages/ProgressionChecker/main";
 import TermPlanner from "./pages/TermPlanner/main";
 import "./App.less";
-import Loading from "./components/Loading/Loading";
+import PageLoading from "./components/PageLoading";
 import "./axios";
+import Header from "./components/Header";
 
 const App = () => {
   const [loading, setLoading] = useState(true);
@@ -17,7 +17,7 @@ const App = () => {
   return (
     <Router>
       {loading ? (
-        <Loading setLoading={setLoading} />
+        <PageLoading setLoading={setLoading} />
       ) : (
         <Routes>
           <Route path="/degree-wizard" element={<DegreeWizard />} />
