+++ frontend/src/assets/helpGifs/additional-options.png
