+++ frontend/src/assets/helpGifs/drag-and-drop.gif
