+++ frontend/src/assets/helpGifs/drag-and-drop.png
