+++ frontend/src/assets/helpGifs/hide-year.gif
