+++ frontend/src/assets/helpGifs/hide-year.png
