+++ frontend/src/components/CourseTag/index.jsx
@@ -1,0 +1,3 @@
+import CourseTag from "./CourseTag";
+
+export default CourseTag;
