+++ frontend/src/components/CourseTag/index.less
