--- frontend/src/components/GlobalColors.js
@@ -1,6 +1,0 @@
-export const lightYellow = "#f9b01e";
-export const lightGrey = "#565652";
-export const darkGrey = "#323739";
-
-export const yellow = "#FAAD14";
-export const purple = "#9254de";
