+++ frontend/src/components/LiquidProgressChart/index.jsx
@@ -1,0 +1,3 @@
+import LiquidProgressChart from "./LiquidProgressChart";
+
+export default LiquidProgressChart;
