+++ frontend/src/components/PageLoading/index.jsx
@@ -1,0 +1,3 @@
+import PageLoading from "./PageLoading";
+
+export default PageLoading;
