+++ frontend/src/components/PlannerCart/PlannerCart.jsx
@@ -1,66 +1,17 @@
 import React, { useState } from "react";
 import {
-  Tooltip, Button, Typography, Popconfirm, Alert,
+  Tooltip, Button, Typography, Alert,
 } from "antd";
 import { useSelector, useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import { CalendarOutlined, DeleteOutlined } from "@ant-design/icons";
-import "./plannerCart.less";
-import { addTab } from "../../reducers/courseTabsSlice";
-import { removeAllCourses, removeCourse } from "../../reducers/plannerSlice";
+import { removeAllCourses } from "../../reducers/plannerSlice";
+import "./index.less";
+import CartCourseCard from "./CartCourseCard";
 
 const { Text, Title } = Typography;
-const CourseCard = ({ code, title, showAlert }) => {
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
-  const [loading, setLoading] = useState(false);
 
-  const confirmDelete = () => {
-    dispatch(removeCourse(code));
-    setLoading(true);
-    setTimeout(() => {
-      showAlert(code);
-      setLoading(false);
-    }, 700);
-  };
-
-  const handleCourseLink = () => {
-    navigate("/course-selector");
-    dispatch(addTab(code));
-  };
-
-  return (
-    <div className="planner-cart-card">
-      <div role="menuitem" onClick={handleCourseLink} style={{ cursor: "pointer" }}>
-        <Text className="text" strong>
-          {code}:{" "}
-        </Text>
-        <Text className="text">{title}</Text>
-      </div>
-      <div className="planner-cart-card-actions">
-        <Popconfirm
-          placement="bottomRight"
-          title="Remove this course from your planner?"
-          onConfirm={confirmDelete}
-          style={{ width: "200px" }}
-          okText="Yes"
-          cancelText="No"
-        >
-          <Tooltip title={`Remove ${code}`}>
-            <Button
-              danger
-              size="small"
-              shape="circle"
-              icon={<DeleteOutlined />}
-              loading={loading}
-            />
-          </Tooltip>
-        </Popconfirm>
-      </div>
-    </div>
-  );
-};
-export const PlannerCart = () => {
+const PlannerCart = () => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
   const courses = useSelector((store) => store.planner.courses);
@@ -110,7 +61,7 @@ export const PlannerCart = () => {
             <div className="planner-cart-content">
               {/* Reversed map to show the most recently added courses first */}
               {Object.keys(courses).reverse().map((courseCode) => (
-                <CourseCard
+                <CartCourseCard
                   code={courseCode}
                   title={courses[courseCode].title}
                   showAlert={showAlert}
@@ -120,7 +71,6 @@ export const PlannerCart = () => {
           ) : (
             <div className="planner-cart-empty-cont">
               <Text className="text">
-                {" "}
                 You have not selected any courses. Find them in our course
                 selector
               </Text>
