+++ frontend/src/components/PlannerCart/index.jsx
@@ -1,0 +1,3 @@
+import PlannerCart from "./PlannerCart";
+
+export default PlannerCart;
