+++ frontend/src/components/RadialBarChart/index.jsx
@@ -1,0 +1,4 @@
+import RadialBarChart from "./RadialBarChart";
+
+// TODO: not currently used
+export default RadialBarChart;
