+++ frontend/src/components/ThemeToggle/index.jsx
@@ -1,0 +1,5 @@
+import ThemeToggle from "./ThemeToggle";
+
+// TODO: currently not being used anywhere..
+// Only some dark mode features have been implemented.
+export default ThemeToggle;
