+++ frontend/src/constants.js
@@ -12,3 +12,10 @@
 export const DATA_STRUCTURE_VERSION = 0;
 
 export const FEEDBACK_LINK = "https://github.com/csesoc/Circles/issues?q=is%3Aissue+is%3Aopen";
+
+// Global colors - currently only being used in LiquidProgressChart
+export const lightYellow = "#f9b01e";
+export const lightGrey = "#565652";
+export const darkGrey = "#323739";
+export const yellow = "#FAAD14";
+export const purple = "#9254de";
