+++ frontend/src/pages/TermPlanner/main.less
@@ -233,7 +233,7 @@ body {
     // background-color: red;
     animation: shake 0.25s;
     transform: none;
-    cursor: url("../../images/notAllowed.ico"), default;
+    cursor: url("../../assets/notAllowed.ico"), default;
   }
 
   @keyframes shake {
